Production/Frontend/requirements.txt
Authlib==0.15.2
boto3==1.15.0
botocore==1.18.18
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
docopt==0.6.2
Flask==1.1.2
hashids==1.3.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pipreqs==0.4.10
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
Werkzeug==1.0.1
yarg==0.1.9
